Split of U2F.php into a Class per File
